Split gui client from base class.
